refactor: Extract utils & component
